requirements.txt
flask-wtf
flask
wtforms
flask-bootstrap4
email-validator